Refactor backend with MVC
