+++ routers/user.py
@@ -2,30 +2,13 @@ from fastapi import *
 from fastapi.responses import JSONResponse
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 from typing import Annotated
-from pydantic import BaseModel, EmailStr
 import os
 import jwt
 import datetime
 from models.user import UserModel
+from views.user import *
 
 router = APIRouter()
-
-class UserInput(BaseModel):
-	name: str
-	email: EmailStr
-	password: str
-
-class UserOutput(BaseModel):
-	id: int
-	name: str
-	email: EmailStr
-
-class UserResponse(BaseModel):
-	data: UserOutput | None = None
-
-class UserSignin(BaseModel):
-	email: EmailStr
-	password: str
 
 security = HTTPBearer(auto_error=False)
 def get_auth_user(credentials: HTTPAuthorizationCredentials= Depends(security)):
